requirements.txt
astroid==3.3.8
blinker==1.9.0
click==8.1.8
colorama==0.4.6
coverage==7.6.10
dill==0.3.9
Flask==3.1.0
greenlet==3.1.1
isort==6.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mccabe==0.7.0
platformdirs==4.3.6
playwright==1.50.0
pyee==12.1.1
pylint==3.3.4
python-dotenv==1.1.0
SQLAlchemy==2.0.39
tomlkit==0.13.2
typing_extensions==4.12.2
Werkzeug==3.1.3
